Finish adding stubs for mc/issues/119
